pyproject.toml
[project]
name = "fintracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.76.0",
    "google-generativeai>=0.3.0",
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
    "openai>=2.15.0",
    "pandas>=2.0.0",
    "pandas-stubs~=2.3.3",
    "pdfplumber>=0.10.0",
    "plotly>=5.18.0",
    "seaborn>=0.13.2",
    "streamlit>=1.30.0",
]
